move db key functions to helper functions
