+++ src/test/scala/vsys/blockchain/state/contract/vswap/ExecuteVSwapValidDiffTest.scala
@@ -3,7 +3,6 @@ package vsys.blockchain.state.contract.vswap
 import org.scalacheck.{Gen, Shrink}
 import org.scalatest.prop.{GeneratorDrivenPropertyChecks, PropertyChecks}
 import org.scalatest.{Matchers, PropSpec}
-import vsys.account.ContractAccount.tokenIdFromBytes
 import vsys.blockchain.block.TestBlock
 import vsys.blockchain.contract.{DataEntry, DataType}
 import vsys.blockchain.contract.token.{SystemContractGen, TokenContractGen}
@@ -12,8 +11,7 @@ import vsys.blockchain.state.diffs._
 import vsys.blockchain.transaction.contract._
 import vsys.blockchain.state._
 import vsys.blockchain.transaction.{GenesisTransaction, TransactionGen, TransactionStatus}
-import com.google.common.primitives.{Bytes, Ints, Longs}
-import vsys.account.PublicKeyAccount
+import com.google.common.primitives.Longs
 
 class ExecuteVSwapValidDiffTest extends PropSpec
   with PropertyChecks
@@ -26,49 +24,6 @@ class ExecuteVSwapValidDiffTest extends PropSpec
   with VSwapFunctionHelperGen {
 
   private implicit def noShrink[A]: Shrink[A] = Shrink(_ => Stream.empty)
-
-  def getContractTokenBalanceKeys(tokenAContractId: Array[Byte], tokenBContractId: Array[Byte], liquidityContractId: Array[Byte], vSwapContractId: Array[Byte]): (ByteStr, ByteStr, ByteStr) = {
-    val tokenAId = tokenIdFromBytes(tokenAContractId, Ints.toByteArray(0)).explicitGet()
-    val tokenBId = tokenIdFromBytes(tokenBContractId, Ints.toByteArray(0)).explicitGet()
-    val liquidityTokenId = tokenIdFromBytes(liquidityContractId, Ints.toByteArray(0)).explicitGet()
-
-    val contractTokenABalanceKey = ByteStr(Bytes.concat(tokenAId.arr, vSwapContractId))
-    val contractTokenBBalanceKey = ByteStr(Bytes.concat(tokenBId.arr, vSwapContractId))
-    val contractLiquidityBalanceKey = ByteStr(Bytes.concat(liquidityTokenId.arr, vSwapContractId))
-
-    (contractTokenABalanceKey, contractTokenBBalanceKey, contractLiquidityBalanceKey)
-  }
-
-  def getUserTokenBalanceKeys(tokenAContractId: Array[Byte], tokenBContractId: Array[Byte], liquidityContractId: Array[Byte], user: PublicKeyAccount): (ByteStr, ByteStr, ByteStr) = {
-    val tokenAId = tokenIdFromBytes(tokenAContractId, Ints.toByteArray(0)).explicitGet()
-    val tokenBId = tokenIdFromBytes(tokenBContractId, Ints.toByteArray(0)).explicitGet()
-    val liquidityTokenId = tokenIdFromBytes(liquidityContractId, Ints.toByteArray(0)).explicitGet()
-
-    val masterTokenABalanceKey = ByteStr(Bytes.concat(tokenAId.arr, user.toAddress.bytes.arr))
-    val masterTokenBBalanceKey = ByteStr(Bytes.concat(tokenBId.arr, user.toAddress.bytes.arr))
-    val masterLiquidityBalanceKey = ByteStr(Bytes.concat(liquidityTokenId.arr, user.toAddress.bytes.arr))
-
-    (masterTokenABalanceKey, masterTokenBBalanceKey, masterLiquidityBalanceKey)
-  }
-
-  def getSwapContractStateVarKeys(vSwapContractId: Array[Byte]): (ByteStr, ByteStr, ByteStr, ByteStr, ByteStr, ByteStr) = {
-    val swapStatusKey = ByteStr(Bytes.concat(vSwapContractId, Array(4.toByte)))
-    val minimumLiquidityKey = ByteStr(Bytes.concat(vSwapContractId, Array(5.toByte)))
-    val tokenAReservedKey = ByteStr(Bytes.concat(vSwapContractId, Array(6.toByte)))
-    val tokenBReservedKey = ByteStr(Bytes.concat(vSwapContractId, Array(7.toByte)))
-    val totalSupplyKey = ByteStr(Bytes.concat(vSwapContractId, Array(8.toByte)))
-    val liquidityTokenLeft = ByteStr(Bytes.concat(vSwapContractId, Array(9.toByte)))
-
-    (swapStatusKey, minimumLiquidityKey, tokenAReservedKey, tokenBReservedKey, totalSupplyKey, liquidityTokenLeft)
-  }
-
-  def getSwapContractStateMapKeys(vSwapContractId: Array[Byte], user: PublicKeyAccount): (ByteStr, ByteStr, ByteStr) = {
-    val stateMapTokenABalanceKey = ByteStr(Bytes.concat(vSwapContractId, Array(0.toByte), DataEntry(user.toAddress.bytes.arr, DataType.Address).bytes))
-    val stateMapTokenBBalanceKey = ByteStr(Bytes.concat(vSwapContractId, Array(1.toByte), DataEntry(user.toAddress.bytes.arr, DataType.Address).bytes))
-    val stateMapLiquidityTokenBalanceKey = ByteStr(Bytes.concat(vSwapContractId, Array(2.toByte), DataEntry(user.toAddress.bytes.arr, DataType.Address).bytes))
-
-    (stateMapTokenABalanceKey, stateMapTokenBBalanceKey, stateMapLiquidityTokenBalanceKey)
-  }
 
   val preconditionsAndVSwapDepositAndWithdrawLiquidityTokens: Gen[(GenesisTransaction, GenesisTransaction, RegisterContractTransaction,
     RegisterContractTransaction, RegisterContractTransaction, RegisterContractTransaction, ExecuteContractFunctionTransaction,
